requirements.txt
pytz==2017.3
python-dateutil==2.6.1
requests==2.18.4
uwsgi
flask
